requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0
supabase==2.0.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
regex==2023.10.3
email-validator==2.1.0
python-dateutil==2.8.2
